src/third_party/blink/web_tests/external/wpt/css/css-overscroll-behavior/parsing/overscroll-behavior-valid-expected.txt

This is a testharness.js-based test.
FAIL e.style['overscroll-behavior'] = "contain" should set the property value assert_equals: serialization should be canonical expected "contain" but got "contain contain"
FAIL e.style['overscroll-behavior'] = "none" should set the property value assert_equals: serialization should be canonical expected "none" but got "none none"
FAIL e.style['overscroll-behavior'] = "auto" should set the property value assert_equals: serialization should be canonical expected "auto" but got "auto auto"
PASS e.style['overscroll-behavior'] = "contain none" should set the property value
PASS e.style['overscroll-behavior'] = "none auto" should set the property value
PASS e.style['overscroll-behavior'] = "auto contain" should set the property value
FAIL e.style['overscroll-behavior'] = "contain contain" should set the property value assert_equals: serialization should be canonical expected "contain" but got "contain contain"
FAIL e.style['overscroll-behavior'] = "none none" should set the property value assert_equals: serialization should be canonical expected "none" but got "none none"
FAIL e.style['overscroll-behavior'] = "auto auto" should set the property value assert_equals: serialization should be canonical expected "auto" but got "auto auto"
PASS e.style['overscroll-behavior-x'] = "contain" should set the property value
PASS e.style['overscroll-behavior-x'] = "none" should set the property value
PASS e.style['overscroll-behavior-x'] = "auto" should set the property value
PASS e.style['overscroll-behavior-y'] = "contain" should set the property value
PASS e.style['overscroll-behavior-y'] = "none" should set the property value
PASS e.style['overscroll-behavior-y'] = "auto" should set the property value
Harness: the test ran to completion.
